Move staging entities to new staging namespace
