+++ api/app/src/Entity/Staging/StagingDeputyship.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OPG\Digideps\Backend\Entity;
+namespace OPG\Digideps\Backend\Entity\Staging;
 
 use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as JMS;
