+++ api/app/src/v2/Registration/DeputyshipProcessing/DeputyshipsCSVLoader.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace OPG\Digideps\Backend\v2\Registration\DeputyshipProcessing;
 
-use OPG\Digideps\Backend\Entity\StagingDeputyship;
-use OPG\Digideps\Backend\v2\CSV\CSVChunkerFactory;
 use Doctrine\ORM\EntityManagerInterface;
 use League\Csv\Exception as CSVException;
 use League\Csv\UnavailableStream;
+use OPG\Digideps\Backend\Entity\Staging\StagingDeputyship;
+use OPG\Digideps\Backend\v2\CSV\CSVChunkerFactory;
 use Psr\Log\LoggerInterface;
 
 class DeputyshipsCSVLoader
@@ -26,8 +26,9 @@ class DeputyshipsCSVLoader
             $chunker = $this->chunkerFactory->create($fileLocation, StagingDeputyship::class);
 
             // delete records from staging table
+            $qualified = StagingDeputyship::class;
             $this->em->beginTransaction();
-            $this->em->createQuery('DELETE FROM OPG\Digideps\Backend\Entity\StagingDeputyship sd')->execute();
+            $this->em->createQuery("DELETE FROM {$qualified} sd")->execute();
             $this->em->flush();
             $this->em->commit();
