+++ api/app/tests/Unit/Factory/StagingSelectedCandidateFactoryTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\OPG\Digideps\Backend\Unit\Factory;
 
-use OPG\Digideps\Backend\Entity\StagingDeputyship;
+use OPG\Digideps\Backend\Entity\Staging\StagingDeputyship;
 use OPG\Digideps\Backend\Factory\StagingSelectedCandidateFactory;
 use PHPUnit\Framework\TestCase;
