+++ api/app/tests/Unit/v2/Registration/DeputyshipProcessing/CourtOrderAndDeputyCandidatesFactoryTest.php
@@ -6,8 +6,8 @@ namespace Tests\OPG\Digideps\Backend\Unit\v2\Registration\DeputyshipProcessing;
 
 use OPG\Digideps\Backend\Domain\Deputyship\DeputyshipProcessingLookupCache;
 use OPG\Digideps\Backend\Entity\CourtOrderDeputy;
-use OPG\Digideps\Backend\Entity\StagingDeputyship;
-use OPG\Digideps\Backend\Entity\StagingSelectedCandidate;
+use OPG\Digideps\Backend\Entity\Staging\StagingDeputyship;
+use OPG\Digideps\Backend\Entity\Staging\StagingSelectedCandidate;
 use OPG\Digideps\Backend\Factory\StagingSelectedCandidateFactory;
 use OPG\Digideps\Backend\Repository\CourtOrderDeputyRepository;
 use OPG\Digideps\Backend\v2\Registration\DeputyshipProcessing\CourtOrderAndDeputyCandidatesFactory;
